+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-23 14:02:51
-// Commit: test(ui): add unit tests for ui business logic
+// Updated at: 2026-04-23 21:43:40
+// Commit: fix(ui): fix memory leak by cancelling ui subscription
 
 class uiComponentState {
-  final String commitRef = '589';
-  final String lastMessage = 'test(ui): add unit tests for ui business logic';
-  final DateTime timestamp = DateTime.parse('2026-04-23T14:02:51+05:00');
+  final String commitRef = '618';
+  final String lastMessage = 'fix(ui): fix memory leak by cancelling ui subscription';
+  final DateTime timestamp = DateTime.parse('2026-04-23T21:43:40+05:00');
 }